ch: refactor `Reader` and `Writer` utilities, switch to file-scoped namespaces, and improve code consistency across `Shared` project
